feat: ajusta estrutura de pastas e implementa props hideCodeTab e hideJSONTab
