+++ jsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./resources/*"]
     }
   },
   "exclude": ["node_modules", "dist"]
