+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@sysvale/form-builder",
-  "version": "0.5.0",
+  "version": "0.6.0",
   "private": false,
   "type": "module",
   "main": "dist/form-builder.umd.js",
