+++ resources/App.vue
