+++ resources/components/SegmentedControl.vue
