+++ resources/index.js
