+++ resources/pt_BR.js
